use atomic pointers to make BPF Send and Sync (#79)
Currently, the BPF struct is marked !Send and !Sync due to our use of bare pointers.

By using the AtomicPtr type, we now make BPF both Send and Sync. This is particularly useful if we want to use BPF within a multithreaded async executor, like Tokio. I have also fixed the llvm 10 builds in Travis by updating the repository name to reflect changes on llvm repo.

No non-obvious implications. Changes are internal to the structures and does not change the surface API.