build/ci.sh - add missing dependencies and make it slightly more user-friendly
This script apparently works well on CI however using it on a
developer machine wasn't a pleasant experience. Some dependencies were
missing (while running on a fresh debian buster box), some variables not set.
Also, the second and consequent runs were failing because of the
leftovers. This commit tries to improve things a bit.